requirements.txt
# Flask and Web Server
flask==3.0.0
flask-cors==4.0.0

# Machine Learning
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2

# Gemini AI
google-genai==1.10.0

# Environment Variables
python-dotenv==1.0.0

# Hardware Integration (Serial Communication)
pyserial==3.5

# Optional: For better performance
scipy==1.11.4
